use mongodb session n storage to become stateless
use store if available, otherwise use local memory

initial work on serverless deployment, got rid of apollo optics
