+++ server/.gitignore
@@ -1,2 +1,3 @@
 artifacts
+libLambda/
 node_modules/
