+++ server/package.json
@@ -16,51 +16,33 @@
     "deploySetup": "./.scripts/deploySetup.sh",
     "deploy": "./.scripts/deploy.sh",
     "heroku-postbuild": "npm install",
-    "build": "BABEL_ENV=development NODE_ENV=production babel src --out-dir lib --watch"
+    "build": "BABEL_ENV=development NODE_ENV=production babel src --out-dir lib --watch",
+    "build:lambda": "BABEL_ENV=lambda NODE_ENV=production babel src --out-dir libLambda --watch",
+    "start:lambda": "NODE_PATH='libLambda' node libLambda/startup-serverless.js"
   },
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "babel-cli": "^6.24.1",
-    "babel-core": "^6.23.1",
-    "babel-eslint": "^7.2.3",
-    "babel-preset-es2015": "^6.22.0",
-    "babel-preset-stage-2": "^6.22.0",
+    "aws-serverless-express": "^3.0.2",
     "bcrypt": "^1.0.2",
     "body-parser": "^1.17.1",
     "connect-mongo": "^1.3.2",
     "cors": "^2.8.1",
-    "eslint-plugin-node": "^4.2.2",
-    "eslint-plugin-promise": "^3.5.0",
-    "eslint-plugin-standard": "^3.0.1",
     "express": "^4.15.2",
     "express-session": "^1.15.2",
     "graphql-server-express": "^0.6.0",
+    "graphql-tools": "^0.10.1",
     "lodash": "^4.17.4",
     "moment": "^2.17.1",
     "mongodb": "^2.2.24",
     "node-fetch": "^1.6.3",
     "node-schedule": "^1.2.4",
     "nodemailer": "^4.0.0",
-    "optics-agent": "^1.1.1",
     "passport": "^0.3.2",
     "passport-local": "^1.0.0",
     "urlencode": "^1.1.0",
-    "babel-plugin-transform-async-to-generator": "^6.22.0",
-    "babel-preset-flow": "^6.23.0",
-    "casual": "^1.5.12",
-    "codecov": "^2.1.0",
-    "eslint": "^3.18.0",
-    "eslint-config-airbnb": "^14.1.0",
-    "eslint-plugin-import": "^2.2.0",
-    "eslint-plugin-jsx-a11y": "^4.0.0",
-    "eslint-plugin-react": "^6.10.0",
-    "flow-bin": "^0.40.0",
-    "graphql-tools": "^0.10.1",
-    "jest": "^22.0.4",
-    "nodemon": "^1.11.0",
-    "standard": "^10.0.2",
-    "testdouble": "^2.1.2"
+    "babel-core": "^6.26.0",
+    "babel-polyfill": "^6.26.0"
   },
   "standard": {
     "parser": "babel-eslint",
@@ -88,11 +70,25 @@
   "devDependencies": {
     "apollo-client": "^1.9.3",
     "apollo-test-utils-with-context": "^0.3.5",
+    "babel-cli": "^6.24.1",
+    "babel-eslint": "^7.2.3",
     "babel-plugin-module-rewrite-with-root": "^0.3.0",
+    "babel-plugin-transform-async-to-generator": "^6.22.0",
+    "babel-preset-env": "^1.6.1",
+    "babel-preset-es2015": "^6.22.0",
+    "babel-preset-flow": "^6.23.0",
+    "babel-preset-stage-2": "^6.22.0",
+    "casual": "^1.5.12",
+    "codecov": "^2.1.0",
+    "flow-bin": "^0.40.0",
     "graphql": "^0.11.7",
     "graphql-tag": "^2.6.1",
+    "jest": "^22.0.4",
     "module-resolver": "^1.0.0",
+    "nodemon": "^1.11.0",
     "snazzy": "^7.0.0",
+    "standard": "^10.0.2",
+    "testdouble": "^2.1.2",
     "thenify": "^3.3.0",
     "tingodb-promise": "0.0.2"
   }
