+++ server/src/startup-serverless.js
@@ -1,0 +1,7 @@
+require('babel-core/register')
+require('babel-polyfill')
+const awsServerlessExpress = require('aws-serverless-express')
+const { app } = require('./app')
+const server = awsServerlessExpress.createServer(app)
+
+exports.handler = (event, context) => awsServerlessExpress.proxy(server, event, context)
